Add files via upload
These two pieces of code are from a group project where we used US Census data. We trained our data to see if we could correctly predict people's income based on all other factors, or in the case of these two files, based on two columns of data from the dataset.